Java Basics 2 finished

+++ Java-Basics-2/src/com/ss/jb2/one/Assignment1.java
@@ -11,17 +11,18 @@ package com.ss.jb2.one;
 public class Assignment1 {
 
 	/**
-	 * @param args
+	 * @param args takes in doubles
 	 * Take multiple values from the command line
 	 * and show the result of adding them
 	 */
 	public static void main(String[] args) {
 
-		float sum = 0;
+		double sum = 0;
 		
+		// parse command line arguments and add to sum
 		for (int i = 0; i < args.length; i++) {
 			try {
-				sum += Float.parseFloat(args[i]);
+				sum += Double.parseDouble(args[i]);
 			}
 			catch (NumberFormatException e) {
 				System.out.println("Error: must enter valid numbers");
